+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [1.1.3] Dev - 2020-05-25
+## [1.1.3] Dev - 2020-05-9
 ### Added
 - New configurable setting to enable/disable escaped [{braces}] (e.g. for creating documentation). Default is enabled.
 - You can now escape tags so they are not processed by wrapping them in [{brackets}]. {{double-braces}} are no longer supported.
+- New {-} tag (soft hyphen)
 - New {profilefullname} tag.
 - New {ifloggedinas}{/ifloggedinas} tags.
 - New {ifnotloggedinas}{/ifnotloggedinas} tags.